// ui/src/brain/data-contracts.d.ts
export interface ProductAttribute {
    /** Name */
    name: string;
    /** Value */
    value: string;
}
// ...
export interface ProductImage {
    /** Url */
    url: string;
    /** Alt */
    alt?: string | null;
}
/** ProductPrice */
export interface ProductPrice {
    /** Amount Cents */
    amount_cents: number;
    /** Amount Float */
    amount_float: number;
    /** Formatted */
    formatted: string;
    /** Currency Code */
    currency_code: string;
}
/** ProductResponse */
export interface ProductResponse {
    /** Id */
    id: string;
    /** Name */
    name: string;
    /** Code */
    code: string;
    /** Description */
    description?: string | null;
    /** Image Url */
    image_url?: string | null;
    price?: ProductPrice | null;
    /**
     * Images
     * @default []
     */
    images?: ProductImage[];
    /**
     * Attributes
     * @default []
     */
    attributes?: ProductAttribute[];
    /** Category */
    category?: string | null;
    /**
     * Available
     * @default true
     */
    available?: boolean;
}